+++ tests/test_context/test_requires_context_io_result/test_context_io_result.py
@@ -4,15 +4,16 @@ import pytest
 
 from returns.context import ContextIOResult, RequiresContextIOResult
 from returns.io import IOFailure, IOSuccess
-from returns.primitives.container import (
+from returns.primitives.exceptions import ImmutableStateError
+from returns.primitives.interfaces import (
     Altable,
     Bindable,
     Fixable,
     Mappable,
     Rescueable,
+    Unitable,
     Unwrapable,
 )
-from returns.primitives.exceptions import ImmutableStateError
 from returns.result import Failure, Success
 
 
@@ -34,6 +35,7 @@ from returns.result import Failure, Success
     Unwrapable,
     Altable,
     Fixable,
+    Unitable,
 ])
 def test_protocols(container, protocol):
     """Ensures that RequiresContext has all the right protocols."""